Split manager
