+++ src/Model/QueryManagerTrait.php
@@ -1,0 +1,37 @@
+<?php
+
+/*
+ * This file is part of the Sonata Project package.
+ *
+ * (c) Thomas Rabaix <thomas.rabaix@sonata-project.org>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sonata\CoreBundle\Model;
+
+trait QueryManagerTrait
+{
+    public function findAll()
+    {
+        return $this->getRepository()->findAll();
+    }
+
+    public function findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
+    {
+        return $this->getRepository()->findBy($criteria, $orderBy, $limit, $offset);
+    }
+
+    public function findOneBy(array $criteria, array $orderBy = null)
+    {
+        return $this->getRepository()->findOneBy($criteria, $orderBy);
+    }
+
+    public function find($id)
+    {
+        return $this->getRepository()->find($id);
+    }
+
+    abstract protected function getRepository();
+}
